Rename operation tests
